Add the option to define gather frequency per measurement
Until now, each device configures a gather and filter update frequency
that applies to all the measurements configured to that device.
With this change we allow to define a particular gather/filter update
frecuency for each measurement.
If one of this values is not defined, it will take the device one.

A gather/filter update frequency defined in one measurement will apply to
all devices with that measurement.
Currently it is not supported to have different gather/filter update
frequency values for the same measurement in different devices.
If this is needed, it could be achieved duplicating the measurement with
different frequency values.

This change has touched the core of SNMP collector therefore a lot of
files have been modified.

The main changes are:
 - adding those new values to the ORM (did in the previous commit)
 - show those values in the UI (previos commits)
 - move the gathering loop from the device to the measurements
 - refactor how statistics work

The core logic to connecto to the devices now resides in the
measurement. It gather data, execute the snmp get, update filters, etc.
Each device has its own goroutine and each measurement for each device
has also its own gorotuine.
Those goroutines are always running for each device, so they can collect
stats even when the device is disabled.
Signals sent from the UI are handled first by the device and then passed
to the measurements.

The same logic for "Active" and "Connected" is done in the measurements,
as it is in the device.

Logging has been moved to use WithFields from logrus. Now its more
structured and easy to parse.

Code has been simplified while possible.

Several other minor bugs have been found and fixed.

A new struct ConnectionParams now stores all the configuration needed to
establish a new snmp connection using goSNMP.
This struct have a function to verify if all parameters are valid.

Filter update now has its own ticker, making easier the change if we
want to move to a definition based on time instead of number of gather
loops.

Sequential gathering is preserved. A shared lock betweek measurement
goroutines forces that only one measurement could gather data at the
same time.
If some measurements take a long time to be processed, could impact in
the frequency of other measurements. This could lead to weird graphics
in visualization.

Fixed some data races found in the bus implementation.

Measurement examples has been adapted to the new syntaxis.

Moved the example of MaxOids to a test. It now test that small values of
MaxOids could lead to invalied queries while getting sysinfo or making
queries from the device.

Enhanced runtime list devices view with new columns that shows the
gather/send stats frequency

Enhanced deview detailed view with:
- Renamed view params on device stats that with this new logic doesn't
have any effect: remaining loops on filterFreq, backend sent duration
and renamed measurement stats to series
- Added new panel with measurement statistics that is referred to
selected measurement on table. Added new statistics to show the user the
start/next time of gather/filter