Leave ABRT signal for debbuging
ABRT signal is used in golang to dump the current goroutines before
exiting.
It is useful to debug possible problems.

Do not leak device bus goroutines

The deviceControlBus goroutine was not being closed when device was
terminated, leaking goroutines, one per device, in each reload config.

Reload config generate an "Exit" signal propagated to devices and from
them to its measurements.
With this signal, devices does not wait for measurement goroutines to
exit and when the device exits it will close the deviceControlBus (and
its internal channels).

If a measurement goroutine is busy doing some processing, the busNode
channel will be closed but the measurement will be able to read the
stored value in that closed channel, so it will eventually exit
correctly.

Anyhow, we add another layer of protection, checking if the channel is
closed while reading from it.
If the channel is closed, the measurement goroutine will exit.